==> 14_Info_timestep_Reconstruction/3_ancestralInfoNull.R <==
# Wrapper Scrit to calculate information by sampling 
# This one is modified to add when a Null strategy is used.

library(MASS)
library(ggplot2)

# Go to the directory and load the data generated by Strategies_mapping.R
# Change this directory to folder with your files as needed
setwd("../14_Info_timestep_Reconstruction/")
rm(list=ls())#empty the workspace
source('localInfo.R')

# Read the data that has the strategies and different points in time
load('stratMapp2.RData')  

# Read the data base of acoustic features for all birds
dataAll=read.table(file="Acoustic_alldata_postDFA.csv", header=T, sep=",",dec=".")

# Global Variables - user specified
nboot <- 30         # Repeat info calculation 30 times at each time point

# Other static variables
Strategies <- data$AcousticClust
Strategies[which(Strategies==1)] = "AC"
Strategies[which(Strategies==2)] = "DK"
Strategies[which(Strategies==3)] = "SF"
Strategies[which(Strategies==4)] = "SS"
Strategies[which(Strategies==5)] = "RS"
Strategies[which(Strategies==6)] = "IS"

nstrat <- length(unique(Strategies))
specInd <- list()
specPerStrat <- array(0, dim=nstrat) # Species per strategy
for (is in 1:nstrat) {
  specInd[[is]] <- which(Strategies == rownames(Q)[is]) 
  specPerStrat[is] <- length(specInd[[is]])
}


# Make space for Info and SE Info
ntimes <- length(tVals)
ancInfo <- array(0, dim=ntimes)
ancInfoSD <- array(0, dim=ntimes)
ancInfoBi <- array(0, dim=ntimes)
ancInfoBiSD <- array(0, dim=ntimes)
ancInfoNull <- array(0, dim=ntimes)
ancInfoSDNull <- array(0, dim=ntimes)
ancInfoBiNull <- array(0, dim=ntimes)
ancInfoBiSDNull <- array(0, dim=ntimes)

nspecies <- array(0, dim=ntimes)

# Breaks for histogram
breaks <- seq(from=0, to=1, by=0.05)
histLI <- array(0, dim=c(ntimes, length(breaks)-1))
nbadboot <- 0

for (it in 1:ntimes) {
  infoVal <- array(0, nboot)
  infoValBi <- array(0, nboot)
  infoValNull <- array(0, nboot)
  infoValBiNull <- array(0, nboot)
  ns <- xVals[[it]][[1]]   # this is the number of "species" at this time point
  nspecies[it] <- ns
  ib <- 1
  while (ib <= nboot) {
    # Generate ns strategies - ss = species strategies sp = species
    ss <- array(0, dim=ns)
    sp <- array(0, dim=ns)
    ssNull <- array(0, dim=ns)
    spNull <- array(0, dim=ns)
    randunif <- runif(ns) 
    for (is in 1:ns) {
      cumP <- cumsum(xVals[[it]][[is+1]])
      ss[is] <- which(randunif[is] < cumP)[1]
    }
    
    badboot <-  FALSE

    # find a strategy for the Null simulation
    indStratNull <- which(specPerStrat >= ns)
    if (length(indStratNull) == 0) {
      infoNullFlg <-  FALSE
    }
    else {
      if (length(indStratNull) == 1) {
        stratNull <- indStratNull
      }
      else {
        stratNull <- sample(indStratNull, size = 1, replace = FALSE)
      }
      infoNullFlg <- TRUE
      for (is in 1:ns) {
        ssNull[is] <- stratNull
      }
    }
    
    # for each species strategy find a random species
    for (istrat in 1:nstrat) {
      nspec <- sum(ss == istrat)
      if ( nspec > length(specInd[[istrat]]) ) {
        print(sprintf('At time slice %d, Strat %s: Asking %d sp but only %d available', it, rownames(Q)[istrat], nspec, length(specInd[[istrat]]) ))
        badboot <-  TRUE
        break
      }

      sp[ss==istrat] <-  sample(specInd[[istrat]], size = nspec, replace = FALSE)

      if (infoNullFlg == TRUE) {
        nspecNull <- sum(ssNull == istrat)    
        spNull[ssNull==istrat] <-  sample(specInd[[istrat]], size = nspecNull, replace = FALSE)
      }

    }
    if (badboot) {
      nbadboot <- nbadboot + 1
      if ( nbadboot > 100 ) {
        break
      }
      next
    }

    # My random sample of species
    spind <-  NULL
    for (is in 1:ns) {
      spind <-  c(spind, which(dataAll$Species == data$Species[sp[is]]))
    }
    
    # Calculate Information for subset
    dataSubset <-  dataAll[spind,]
    info <- localInfo(dataSubset)
    
    # Store mean
    infoVal[ib] <- sum(info$Full, na.rm=TRUE)/length(info$Full)
    infoValBi[ib] <- sum(info$Binary, na.rm=TRUE)/length(info$Binary)
    
    # Store histogram
    histval <- hist(info$Full/log2(ns), breaks = breaks, plot=FALSE)
    histLI[it,] <- histLI[it,] + histval$counts
    
    # Repeat for Null
    if (infoNullFlg == TRUE) {
      spind <-  NULL
      for (is in 1:ns) {
        spind <-  c(spind, which(dataAll$Species == data$Species[spNull[is]]))
      }
      # Calculate Information for subset
      dataSubset <-  dataAll[spind,]
      infoNull <- localInfo(dataSubset)
      
      # Store mean
      infoValNull[ib] <- sum(infoNull$Full, na.rm=TRUE)/length(infoNull$Full)
      infoValBiNull[ib] <- sum(infoNull$Binary, na.rm=TRUE)/length(infoNull$Binary)
    }
    else {
      infoValNull[ib] <- NA
      infoValBiNull[ib] <- NA
    }
    ib <- ib + 1
  }
  ancInfo[it] <- mean(infoVal, na.rm=TRUE)
  ancInfoSD[it] <- sd(infoVal, na.rm=TRUE)
  ancInfoBi[it] <- mean(infoValBi, na.rm=TRUE)
  ancInfoBiSD[it] <- sd(infoValBi, na.rm=TRUE)
  
  if (infoNullFlg == TRUE) {
  ancInfoNull[it] <- mean(infoValNull, na.rm=TRUE)
  ancInfoSDNull[it] <- sd(infoValNull, na.rm=TRUE)
  ancInfoBiNull[it] <- mean(infoValBiNull, na.rm=TRUE)
  ancInfoBiSDNull[it] <- sd(infoValBiNull, na.rm=TRUE)
  }
}

# Calculate info at current time.
infoNow <- localInfo(dataAll)
tVals[ntimes+1] <- tVals[ntimes]+tVals[1]
nspecies[ntimes+1] <- length(unique(dataAll$Species))
tVals <- tVals - tVals[ntimes+1]   # Set current time to zero
ancInfo[ntimes+1] <- sum(infoNow$Full, na.rm=TRUE)/length(infoNow$Full)
ancInfoSD[ntimes+1] <- 0
ancInfoBi[ntimes+1] <- sum(infoNow$Binary, na.rm=TRUE)/length(infoNow$Binary)
ancInfoBiSD[ntimes+1] <- 0
ancInfoNull[ntimes+1] <- 0
ancInfoSDNull[ntimes+1] <- 0
ancInfoBiNull[ntimes+1] <- 0
ancInfoBiSDNull[ntimes+1] <- 0


dataPlot <- data.frame(tVals, ancInfo, ancInfoSD, ancInfoBi, ancInfoBiSD, ancInfoNull, ancInfoSDNull, ancInfoBiNull, ancInfoBiSDNull)

# Print Info
p<- ggplot(dataPlot, aes(x=tVals, y=ancInfo) ) + 
  geom_line() +
  geom_point()+
  geom_errorbar(aes(ymin=ancInfo-ancInfoSD, ymax=ancInfo+ancInfoSD), width=.2,
                position=position_dodge(0.05))
p+labs(x="Time (mYrs)", y = "Info (Bits)")+ylim(c(0.0, 2.5)) +
  theme_classic()

# Normalized Info
p2<- ggplot(dataPlot, aes(x=tVals, y=100.0*ancInfo/log2(nspecies)) ) + 
  geom_line() +
  geom_point()+
  geom_errorbar(aes(ymin=100.0*(ancInfo-ancInfoSD)/log2(nspecies), ymax=100.0*(ancInfo+ancInfoSD)/log2(nspecies)),
                width=.2, position=position_dodge(0.05))
p2+labs(x="Time (mYrs)", y = "Normalized Info (%)")+ylim(c(0.0, 100.0)) +
  theme_classic()

# Dual Plot
p4 <- ggplot(dataPlot, aes(x=tVals))
  # The normalized info
  p4 <- p4 + geom_line( aes(y=100.0*ancInfo/log2(nspecies), colour='Normalized') ) 
  p4 <- p4 + geom_point(aes(y=100.0*ancInfo/log2(nspecies), colour='Normalized') ) 
  p4 <- p4 + geom_ribbon(aes(ymin=100.0*(ancInfo-ancInfoSD)/log2(nspecies), ymax=100.0*(ancInfo+ancInfoSD)/log2(nspecies), colour="Normalized"),
                alpha=.3, fill = 'blue')
  
  p4 <- p4 + labs(x="Time (mYrs)", y = "Normalized Info (%)")
  
  # The total information on a secondary axis.
  p4 <- p4 + geom_line( aes(y=ancInfo*40, colour = 'Raw'))
  p4 <- p4 + geom_point(aes(y=ancInfo*40, colour = 'Raw')) 
  p4 <- p4 + geom_ribbon( aes(ymin=(ancInfo-ancInfoSD)*40, ymax=(ancInfo+ancInfoSD)*40, colour='Raw'), alpha=.3, fill='red')
  p4 <- p4 + scale_y_continuous(sec.axis = sec_axis(~./40, name =  "Info (Bits)"), limits = c(0,100)) 

  p4 <- p4 + scale_colour_manual(values = c("blue", "red"))
  p4 <- p4 + theme_minimal()
  p4 <- p4 + theme(legend.position = c(0.2, 0.95))
  p4

#ggsave("ancInfo.eps", device = "eps", width = 15, height = 10, units = "cm")

# With control data
#pdf ("SupFig_infoRec.pdf", height = 5, width = 7)
p5<- ggplot(dataPlot,aes(x=tVals, y=ancInfo, colour='Actual')) + 
  geom_line(aes(x=tVals, y=ancInfo, colour='Actual')) +
  geom_point(aes(x=tVals, y=ancInfo, colour = 'Actual'))+
  geom_errorbar(aes(ymin=ancInfo-ancInfoSD, ymax=ancInfo+ancInfoSD, colour='Actual'), width=.2,
                position=position_dodge(0.05)) +
  geom_line(aes(x=tVals, y=ancInfoNull, colour = 'One Strat')) +
  geom_point(aes(x=tVals, y=ancInfoNull, colour = 'One Strat'))+
  geom_errorbar(aes(ymin=ancInfoNull-ancInfoSDNull, ymax=ancInfoNull+ancInfoSDNull, colour = 'One Strat'), width=.2,
                position=position_dodge(0.05)) +
  scale_colour_manual(values = c("red", "black"))
  
  
p5 <- p5+labs(x="Time (mYrs)", y = "Info (Bits)") + ylim(c(0.0, 2.5)) + xlim(c(-21, -5))+
  theme(legend.position = c(0.1, 0.8)) + theme_classic()
p5


p6<- ggplot(dataPlot, aes(x=tVals, y=ancInfo/log2(nspecies), colour='Actual') ) + 
  geom_line(aes(x=tVals, y=ancInfo/log2(nspecies), colour='Actual')) +
  geom_point(aes(x=tVals, y=ancInfo/log2(nspecies), colour = 'Actual'))+
  geom_errorbar(aes(ymin=(ancInfo-ancInfoSD)/log2(nspecies), ymax=(ancInfo+ancInfoSD)/log2(nspecies), colour='Actual'), width=.2,
                position=position_dodge(0.05)) +
  geom_line(aes(x=tVals, y=ancInfoNull/log2(nspecies), colour='One Strat')) +
  geom_point(aes(x=tVals, y=ancInfoNull/log2(nspecies), colour = 'One Strat'))+
  geom_errorbar(aes(ymin=(ancInfoNull-ancInfoSDNull)/log2(nspecies), ymax=(ancInfoNull+ancInfoSDNull)/log2(nspecies), colour='One Strat'), width=.2,
                position=position_dodge(0.05))+
  scale_colour_manual(values = c("red", "black"))

p6 <- p6+labs(x="Time (mYrs)", y = "Normalized Info %") + ylim(c(0.0, 1.0)) +  xlim(c(-21, -5)) +
  theme(legend.position = c(0.1, 0.8)) + theme_classic()
p6
#dev.off()

# save data
save(nspecies, tVals, ancInfo, ancInfoSD, ancInfoBi, ancInfoBiSD, ancInfoNull, ancInfoSDNull, ancInfoBiNull, ancInfoBiSDNull, histLI, infoNow, file='infoRes.RData')
